Move generation of the application graph to the beginning of the migrator
