+++ mule-migration-tool-library/pom.xml
@@ -142,13 +142,6 @@
             <artifactId>snakeyaml</artifactId>
         </dependency>
 
-        <dependency>
-            <groupId>org.jgrapht</groupId>
-            <artifactId>jgrapht-core</artifactId>
-            <version>1.4.0</version>
-        </dependency>
-        
-
         <!-- Test Dependencies -->
         <dependency>
             <groupId>${project.groupId}</groupId>
